refactor(agent-sessions): move termination into agentSession.service (NAN-6598)
Connect sessions keep the row access and the operation emitting layer in one service, so agent
sessions do the same. The row level function is now endAgentSession and terminateAgentSession
is the one that emits the operation and maps the error codes.

Creation moves in a follow up.
